+++ api/models/__init__.py
@@ -1,0 +1,4 @@
+from .models import *
+from .spell import Spell
+from .monster import Monster
+from .monster import MonsterSpell
